Updating to be SDK 5.x version compatible
Our shared test structure references namespaces that no longer exist in
the 5.x release of the SDK.

In addition to the above change, this commit also introduces the
following set of changes:

- Bump minimum PHP to 8.0 in accordance with the PHP SDK requirements.
- Updates README with new recommendations for expected usage of this
  repository.
- Makes some stylistic improvements (spaces over tabs, no ending closing
  tag)
- Makes some LSP sourced recommendations (one class per file, docblocks,
  return types)
