+++ src/DatabaseFeatureRequesterTestBase.php
@@ -1,10 +1,11 @@
 <?php
 
+
 namespace LaunchDarkly\SharedTest;
 
-use LaunchDarkly\FeatureRequester;
 use LaunchDarkly\Impl\Model\FeatureFlag;
 use LaunchDarkly\Impl\Model\Segment;
+use LaunchDarkly\Subsystems\FeatureRequester;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -17,7 +18,7 @@ class DatabaseFeatureRequesterTestBase extends TestCase
     /**
      * Override this method to remove all data from the underlying data store for
      * the specified prefix string.
-     * 
+     *
      * @param string $prefix the key prefix; may be empty/null, in which case we should
      *   use whatever the default prefix is for this database integration.
      */
@@ -29,11 +30,11 @@ class DatabaseFeatureRequesterTestBase extends TestCase
     /**
      * Override this method to create an instance of the feature requester class being
      * tested.
-     * 
+     *
      * @param string $prefix the key prefix; may be empty/null, in which case we should
      *   use whatever the default prefix is for this database integration.
-     * 
-     * @return an implementation instance
+     *
+     * @return FeatureRequester an implementation instance
      */
     protected function makeRequester(?string $prefix): FeatureRequester
     {
@@ -63,7 +64,7 @@ class DatabaseFeatureRequesterTestBase extends TestCase
     /**
      * @dataProvider prefixParameters
      */
-    public function testGetFeature(?string $prefix)
+    public function testGetFeature(?string $prefix): void
     {
         $this->clearExistingData($prefix);
         $fr = $this->makeRequester($prefix);
@@ -83,7 +84,7 @@ class DatabaseFeatureRequesterTestBase extends TestCase
     /**
      * @dataProvider prefixParameters
      */
-    public function testGetMissingFeature(?string $prefix)
+    public function testGetMissingFeature(?string $prefix): void
     {
         $this->clearExistingData($prefix);
         $fr = $this->makeRequester($prefix);
@@ -95,7 +96,7 @@ class DatabaseFeatureRequesterTestBase extends TestCase
     /**
      * @dataProvider prefixParameters
      */
-    public function testGetDeletedFeature(?string $prefix)
+    public function testGetDeletedFeature(?string $prefix): void
     {
         $this->clearExistingData($prefix);
         $fr = $this->makeRequester($prefix);
@@ -113,7 +114,7 @@ class DatabaseFeatureRequesterTestBase extends TestCase
     /**
      * @dataProvider prefixParameters
      */
-    public function testGetAllFeatures(?string $prefix)
+    public function testGetAllFeatures(?string $prefix): void
     {
         $this->clearExistingData($prefix);
         $fr = $this->makeRequester($prefix);
@@ -131,7 +132,7 @@ class DatabaseFeatureRequesterTestBase extends TestCase
         $this->putSerializedItem($prefix, 'features', $flagKey3, $flagVersion, $flagJson3);
 
         $flags = $fr->getAllFeatures();
-        
+
         $this->assertEquals(2, count($flags));
         $flag1 = $flags[$flagKey1];
         $this->assertEquals($flagKey1, $flag1->getKey());
@@ -144,7 +145,7 @@ class DatabaseFeatureRequesterTestBase extends TestCase
     /**
      * @dataProvider prefixParameters
      */
-    public function testAllFeaturesWithEmptyStore(?string $prefix)
+    public function testAllFeaturesWithEmptyStore(?string $prefix): void
     {
         $this->clearExistingData($prefix);
         $fr = $this->makeRequester($prefix);
@@ -156,7 +157,7 @@ class DatabaseFeatureRequesterTestBase extends TestCase
     /**
      * @dataProvider prefixParameters
      */
-    public function testGetSegment(?string $prefix)
+    public function testGetSegment(?string $prefix): void
     {
         $this->clearExistingData($prefix);
         $fr = $this->makeRequester($prefix);
@@ -175,7 +176,7 @@ class DatabaseFeatureRequesterTestBase extends TestCase
     /**
      * @dataProvider prefixParameters
      */
-    public function testGetMissingSegment(?string $prefix)
+    public function testGetMissingSegment(?string $prefix): void
     {
         $this->clearExistingData($prefix);
         $fr = $this->makeRequester($prefix);
@@ -187,7 +188,7 @@ class DatabaseFeatureRequesterTestBase extends TestCase
     /**
      * @dataProvider prefixParameters
      */
-    public function testGetDeletedSegment(?string $prefix)
+    public function testGetDeletedSegment(?string $prefix): void
     {
         $this->clearExistingData($prefix);
         $fr = $this->makeRequester($prefix);
@@ -202,7 +203,7 @@ class DatabaseFeatureRequesterTestBase extends TestCase
         $this->assertNull($segment);
     }
 
-    public function testPrefixIndependence()
+    public function testPrefixIndependence(): void
     {
         $prefix1 = 'prefix1';
         $prefix2 = 'prefix2';
@@ -226,7 +227,7 @@ class DatabaseFeatureRequesterTestBase extends TestCase
         $this->verifyForPrefix($this->makeRequester($prefix2), $flagKey, $segmentKey, $version2);
     }
 
-    private function setupForPrefix(?string $prefix, string $flagKey, string $segmentKey, int $flagVersion)
+    private function setupForPrefix(?string $prefix, string $flagKey, string $segmentKey, int $flagVersion): void
     {
         $segmentVersion = $flagVersion * 2;
         $this->putSerializedItem($prefix, 'features', $flagKey, $flagVersion,
@@ -235,7 +236,7 @@ class DatabaseFeatureRequesterTestBase extends TestCase
             self::makeSegmentJson($flagKey, $segmentVersion));
     }
 
-    private function verifyForPrefix(FeatureRequester $fr, string $flagKey, string $segmentKey, int $flagVersion)
+    private function verifyForPrefix(FeatureRequester $fr, string $flagKey, string $segmentKey, int $flagVersion): void
     {
         $segmentVersion = $flagVersion * 2;
 
@@ -252,7 +253,10 @@ class DatabaseFeatureRequesterTestBase extends TestCase
         $this->assertEquals($segmentVersion, $segment->getVersion());
     }
 
-    public function prefixParameters()
+    /**
+     * @return array<int,mixed>
+     */
+    public function prefixParameters(): array
     {
         return [
             [ self::TEST_PREFIX ],
@@ -261,7 +265,7 @@ class DatabaseFeatureRequesterTestBase extends TestCase
         ];
     }
 
-    private static function makeFlagJson(string $key, int $version, bool $deleted = false)
+    private static function makeFlagJson(string $key, int $version, bool $deleted = false): string|bool
     {
         return json_encode(array(
             'key' => $key,
@@ -283,7 +287,7 @@ class DatabaseFeatureRequesterTestBase extends TestCase
         ));
     }
 
-    private static function makeSegmentJson(string $key, int $version, bool $deleted = false)
+    private static function makeSegmentJson(string $key, int $version, bool $deleted = false): string|bool
     {
         return json_encode(array(
             'key' => $key,
@@ -296,5 +300,3 @@ class DatabaseFeatureRequesterTestBase extends TestCase
         ));
     }
 }
-
-?>
